faq/views.py
from rest_framework import viewsets
from rest_framework.response import Response
from .models import FAQ
from .serializers import FAQSerializer


class FAQViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = FAQ.objects.all()
    serializer_class = FAQSerializer

    def list(self, request, *args, **kwargs):
        lang = request.query_params.get('lang', 'en')
        faqs = self.queryset.all()
        serializer = FAQSerializer(faqs, many=True, context={'lang': lang})
        return Response(serializer.data)
