requirements.txt
Django==4.2
djangorestframework==3.14.0
django-ckeditor==6.7.0
googletrans==4.0.0-rc1
django-redis==5.3.0
redis==5.0.1
pytest==7.4.3
pytest-django==4.5.2
flake8==6.1.0